Exti handler example using HAL
The PAL events only goes up to 15 (i.e. actual gpios) for the other ones this is how to do it. It was quite confusing to figure out what it wanted as none of these defines are used by any example code. It is actually quite simple though.

To enable the nvic vector use
#define STM32_EXTI16_IS_USED

To enable the interrupt use
extiEnableLine(16, EXTI_MODE_RISING_EDGE | EXTI_MODE_ACTION_INTERRUPT);

To define the interrupt handler use
#define STM32_EXTI16_ISR(pr, line) hw_axiom_pvd_interrupt_handler(pr, line)

The pr bit is cleared before entering the handler and the old value of pr is passed in, along with which line the interrupt is for. There is no checking if the bit was set before calling the handler. So a simple check to check this is the right interrupt is required.
if((pr & EXTI_MASK1(line)) != RESET) { }